server update

+++ app/controllers/SessionsController.php
@@ -26,10 +26,21 @@ class SessionsController extends BaseController
             );
 
             if (count($users) != 1) {
-                Session::delete('username');
-                Auth::clear('default');
+                if (isset($_POST['latitude']) && isset($_POST['longitude'])) {
+                    $this->_render['layout']   = 'json';
+                    $this->_render['template'] = 'mobile';
+                    $this->_render['type']     = 'json';
 
-                return $this->redirect('/logout');
+                    return array(
+                        'data'   => null,
+                        'errors' => 'User not found!'
+                    );
+                } else {
+                    Session::delete('username');
+                    Auth::clear('default');
+
+                    return $this->redirect('/logout');
+                }
             }
             $user = $users->current();
